refactor(agent): 全大写常量收敛到包内 constants.ts
bridge 与 agent 两包把散落 / 重复的全大写常量统一到包内 constants.ts，便于复用与调参。

@meebox/pr-agent-bridge constants.ts：超时（run / chat / probe）、UTF8_ENV、PRAGENT_LOCAL_OUTPUT。
顺带修两处重复 / 魔法值：bridge.ts buildInvocation 内联 utf8 与模块 UTF8_ENV 合一；detect.ts 探测
超时 5000 抽成 DEFAULT_PROBE_TIMEOUT_MS。

@meebox/agent constants.ts：VERDICTS（原 planning/review 各一份重复）、AGENT_FILES、
AGENT_RULES_SUBDIR、EMPTY_FILES、READ_TOOLS、MUTATING_TOOLS、STAGGER_MIN/SPAN_MS、
MAX_PARALLEL_TOOLS、HISTORY_MESSAGE_MAX、HISTORY_BUDGET_CHARS、JUDGE_SYSTEM、
JUDGE_MAX_OUTPUT_TOKENS、DESC_CLAMP、CACHE_BREAK。各源文件改为从 constants 引用；index 公共导出
（AGENT_FILES / AGENT_RULES_SUBDIR / READ_TOOLS / MUTATING_TOOLS）改走 constants，公共 API 不变。

不收两类非「值常量」：step 注册表 REVIEW_STEPS（实例化各 step 类的组合根）、?raw 资源载入对象
PROMPT_TEMPLATES / AGENT_TEMPLATES（各自模块本体）。STEP_LABELS / SUMMARY_SECTIONS 暂留，由后续
i18n 统一项移除。纯结构调整、行为不变。
